[Interface] Modified structure to monolithic build for simplicity
