--- engine/src/bbk_beam/mtRooms/engine/db/IDatabase.java
@@ -1,4 +1,0 @@
-package bbk_beam.mtRooms.engine.db;
-
-public interface IDatabase {
-}
