--- engine/src/bbk_beam/mtRooms/engine/reservation/IOptimisedSearch.java
@@ -1,4 +1,0 @@
-package bbk_beam.mtRooms.engine.reservation;
-
-public interface IOptimisedSearch {
-}
