--- engine/src/bbk_beam/mtRooms/engine/revenue/IInvoiceSender.java
@@ -1,4 +1,0 @@
-package bbk_beam.mtRooms.engine.revenue;
-
-public interface IInvoiceSender {
-}
