+++ src/db/IDatabase.java
@@ -1,0 +1,4 @@
+package bbk_beam.mtRooms.db;
+
+public interface IDatabase {
+}
