+++ src/reservation/IPaymentProcessing.java
@@ -1,0 +1,4 @@
+package bbk_beam.mtRooms.reservation;
+
+public interface IPaymentProcessing {
+}
