+++ src/reservation/IReservationProcess.java
@@ -1,0 +1,4 @@
+package bbk_beam.mtRooms.reservation;
+
+public interface IReservationProcess {
+}
