+++ src/revenue/IInvoiceSender.java
@@ -1,0 +1,4 @@
+package bbk_beam.mtRooms.revenue;
+
+public interface IInvoiceSender {
+}
